refactor(tools): move react tools from legacy/tools/ to agent/tools/
createMessageReactTool / createReactTools are AgentTool factories used
only by createAgentTools. Belong with the rest of the agent tools.

- mv legacy/tools/react.{ts,test.ts} → agent/tools/react.{ts,test.ts}
- update 4 import sites (app.ts, agent/runtime.ts, agent/tools/index.ts,
  legacy/plugins/types.ts)
- remove from legacy/tools/index.ts barrel
- drop stale src/tools/react.ts file-header comment

Net: file moves only, no behavior change.
