+++ src/agent/tools/index.ts
@@ -12,7 +12,7 @@ import { HostFs, SandboxFs, type FsBridge } from "../../sandbox/fs-bridge.js";
 import { SandboxExecutor } from "../../sandbox/executor.js";
 import { type SandboxConfig } from "../../sandbox/config.js";
 import { createWebFetchTool } from "./web.js";
-import { createReactTools, type LazyTransportContext } from "../../legacy/tools/react.js";
+import { createReactTools, type LazyTransportContext } from "./react.js";
 import { createExecTools, ProcessRegistry } from "../../legacy/tools/exec.js";
 import type { AgentRuntime } from "../runtime.js";
 import { RunValidationError } from "../types.js";
