+++ src/legacy/tools/index.ts
@@ -6,9 +6,3 @@ export {
   createExecTools,
 } from "./exec.js";
 export type { ExecToolOptions, ProcessInfo } from "./exec.js";
-
-export {
-  createMessageReactTool,
-  createReactTools,
-} from "./react.js";
-export type { ReactToolContext } from "./react.js";
